+++ demos/petclinic/src/main/java/is/codion/demos/petclinic/ui/PetclinicAppPanel.java
@@ -33,6 +33,7 @@ import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.plugin.flatlaf.intellij.themes.arc.Arc;
 import is.codion.swing.common.ui.component.indicator.ValidIndicatorFactory;
 import is.codion.swing.framework.model.SwingEntityModel;
+import is.codion.swing.framework.ui.EntityApplication;
 import is.codion.swing.framework.ui.EntityApplicationPanel;
 import is.codion.swing.framework.ui.EntityPanel;
 import is.codion.swing.framework.ui.ReferentialIntegrityErrorHandling;
@@ -122,7 +123,7 @@ public final class PetclinicAppPanel extends EntityApplicationPanel<PetclinicApp
 		ReferentialIntegrityErrorHandling.REFERENTIAL_INTEGRITY_ERROR_HANDLING
 						.set(ReferentialIntegrityErrorHandling.DISPLAY_DEPENDENCIES);
 		ValidIndicatorFactory.FACTORY_CLASS.set("is.codion.plugin.flatlaf.indicator.FlatLafValidIndicatorFactory");
-		EntityApplicationPanel.builder(PetclinicAppModel.class, PetclinicAppPanel.class)
+		EntityApplication.builder(PetclinicAppModel.class, PetclinicAppPanel.class)
 						.domain(Petclinic.DOMAIN)
 						.startupDialog(false)
 						.defaultLookAndFeel(Arc.class)