+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityApplicationPanel.java
@@ -24,17 +24,14 @@ import is.codion.common.i18n.Messages;
 import is.codion.common.logging.LoggerProxy;
 import is.codion.common.model.CancelException;
 import is.codion.common.model.preferences.UserPreferences;
-import is.codion.common.observer.Observable;
 import is.codion.common.observer.Observer;
 import is.codion.common.property.PropertyStore;
 import is.codion.common.property.PropertyValue;
 import is.codion.common.resource.MessageBundle;
 import is.codion.common.state.State;
-import is.codion.common.user.User;
 import is.codion.common.version.Version;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.db.EntityConnectionTracer;
-import is.codion.framework.domain.DomainType;
 import is.codion.framework.domain.entity.Entities;
 import is.codion.framework.domain.entity.EntityDefinition;
 import is.codion.framework.domain.entity.EntityType;
@@ -66,7 +63,6 @@ import org.jspecify.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.swing.ImageIcon;
 import javax.swing.JComponent;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
@@ -75,7 +71,6 @@ import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.JTree;
-import javax.swing.LookAndFeel;
 import javax.swing.ScrollPaneConstants;
 import javax.swing.UIManager;
 import javax.swing.WindowConstants;
@@ -107,7 +102,6 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.function.Supplier;
 import java.util.prefs.Preferences;
 import java.util.stream.Stream;
 
@@ -132,7 +126,7 @@ import static javax.swing.UIManager.getLookAndFeel;
 /**
  * A central application panel class.
  * @param <M> the application model type
- * @see #builder(Class, Class)
+ * @see EntityApplication#builder(Class, Class)
  */
 public class EntityApplicationPanel<M extends SwingEntityApplicationModel> extends JPanel {
 
@@ -508,18 +502,6 @@ public class EntityApplicationPanel<M extends SwingEntityApplicationModel> exten
 		}
 
 		return new DefaultTreeModel(root);
-	}
-
-	/**
-	 * @param <M> the application model type
-	 * @param <P> the application panel type
-	 * @param applicationModelClass the application model class
-	 * @param applicationPanelClass the application panel class
-	 * @return a {@link Builder}
-	 */
-	public static <M extends SwingEntityApplicationModel, P extends EntityApplicationPanel<M>> Builder<M, P> builder(
-					Class<M> applicationModelClass, Class<P> applicationPanelClass) {
-		return new DefaultEntityApplicationPanelBuilder<>(applicationModelClass, applicationPanelClass);
 	}
 
 	/**
@@ -1356,225 +1338,5 @@ public class EntityApplicationPanel<M extends SwingEntityApplicationModel> exten
 		 * @see EntityPanel#activated()
 		 */
 		default void activated(EntityPanel entityPanel) {}
-	}
-
-	/**
-	 * Builds a {@link EntityApplicationPanel} and starts the application.
-	 * @param <M> the application model type
-	 * @param <P> the application panel type
-	 * @see EntityApplicationPanel#builder(Class, Class)
-	 * @see #start()
-	 */
-	public interface Builder<M extends SwingEntityApplicationModel, P extends EntityApplicationPanel<M>> {
-
-		/**
-		 * @param domain the domain type
-		 * @return this Builder instance
-		 */
-		Builder<M, P> domain(DomainType domain);
-
-		/**
-		 * <p>Sets the application name, used as frame title and client type identifier when using remote connnections.
-		 * <p>If no application name is set, {@link DomainType#name()} is used or
-		 * {@code applicationPanelClass.getSimpleName()} in case of no domain model.
-		 * @param applicationName the application name
-		 * @return this Builder instance
-		 */
-		Builder<M, P> applicationName(String applicationName);
-
-		/**
-		 * @param applicationIcon the application icon
-		 * @return this Builder instance
-		 */
-		Builder<M, P> applicationIcon(ImageIcon applicationIcon);
-
-		/**
-		 * @param version the application version
-		 * @return this Builder instance
-		 */
-		Builder<M, P> version(Version version);
-
-		/**
-		 * Sets the default look and feel class, used in case no look and feel settings are found in user preferences.
-		 * @param defaultLookAndFeelClass the default look and feel class
-		 * @return this Builder instance
-		 * @see LookAndFeelProvider
-		 */
-		Builder<M, P> defaultLookAndFeel(Class<? extends LookAndFeel> defaultLookAndFeelClass);
-
-		/**
-		 * Sets the default look and feel classname, used in case no look and feel settings are found in user preferences.
-		 * @param defaultLookAndFeelClassName the default look and feel classname
-		 * @return this Builder instance
-		 * @see LookAndFeelProvider
-		 */
-		Builder<M, P> defaultLookAndFeel(String defaultLookAndFeelClassName);
-
-		/**
-		 * Sets the look and feel class, overrides any look and feel settings found in user preferences.
-		 * @param lookAndFeelClass the look and feel class
-		 * @return this Builder instance
-		 * @see LookAndFeelProvider
-		 */
-		Builder<M, P> lookAndFeel(Class<? extends LookAndFeel> lookAndFeelClass);
-
-		/**
-		 * Sets the look and feel classname, overrides any look and feel settings found in user preferences.
-		 * @param lookAndFeelClassName the look and feel classname
-		 * @return this Builder instance
-		 * @see LookAndFeelProvider
-		 */
-		Builder<M, P> lookAndFeel(String lookAndFeelClassName);
-
-		/**
-		 * Overrides {@link #connectionProvider(Function)}
-		 * @param connectionProvider the connection provider
-		 * @return this Builder instance
-		 */
-		Builder<M, P> connectionProvider(EntityConnectionProvider connectionProvider);
-
-		/**
-		 * @param connectionProvider initializes the connection provider, receives the user provided by {@link #user(Supplier)}
-		 * @return this Builder instance
-		 */
-		Builder<M, P> connectionProvider(Function<User, EntityConnectionProvider> connectionProvider);
-
-		/**
-		 * @param applicationModel the application model factory
-		 * @return this Builder instance
-		 */
-		Builder<M, P> applicationModel(Function<EntityConnectionProvider, M> applicationModel);
-
-		/**
-		 * @param applicationPanel the application panel factory
-		 * @return this Builder instance
-		 */
-		Builder<M, P> applicationPanel(Function<M, P> applicationPanel);
-
-		/**
-		 * <p>The {@link User} to use to connect to the database, this user is propagated to {@link #connectionProvider(Function)}.
-		 * <p>If this user is null, {@link #user(Supplier)} is used to fetch a user.
-		 * @param user the application user
-		 * @return this Builder instance
-		 * @see is.codion.framework.model.EntityApplicationModel#USER
-		 */
-		Builder<M, P> user(@Nullable User user);
-
-		/**
-		 * <p>Supplies the {@link User} to use to connect to the database, this user is then propagated to {@link #connectionProvider(Function)}.
-		 * <p>This may be via a login dialog or simply by returning a hardcoded instance.
-		 * <p>Startup is silently cancelled in case the {@link Supplier#get()} throws a {@link CancelException}.
-		 * @param userSupplier supplies the application user, for example via a login dialog
-		 * @return this Builder instance
-		 */
-		Builder<M, P> user(Supplier<User> userSupplier);
-
-		/**
-		 * @param defaultUser the default user credentials to display in a login dialog
-		 * @return this Builder instance
-		 */
-		Builder<M, P> defaultUser(@Nullable User defaultUser);
-
-		/**
-		 * @param saveDefaultUsername true if the username should be saved in user preferences after a successful login
-		 * @return this Builder instance
-		 */
-		Builder<M, P> saveDefaultUsername(boolean saveDefaultUsername);
-
-		/**
-		 * Note that this does not apply when a custom {@link #user(Supplier)} has been specified.
-		 * @param loginPanelSouthComponentSupplier supplies the component to add to the
-		 * {@link BorderLayout#SOUTH} position of the default login panel
-		 * @return this Builder instance
-		 */
-		Builder<M, P> loginPanelSouthComponent(Supplier<JComponent> loginPanelSouthComponentSupplier);
-
-		/**
-		 * Runs before the application is started, but after Look and Feel initialization.
-		 * Throw {@link CancelException} in order to cancel the application startup.
-		 * @param beforeApplicationStarted run before the application is started
-		 * @return this Builder instance
-		 */
-		Builder<M, P> beforeApplicationStarted(@Nullable Runnable beforeApplicationStarted);
-
-		/**
-		 * @param onApplicationStarted called after a successful application start
-		 * @return this Builder instance
-		 */
-		Builder<M, P> onApplicationStarted(@Nullable Consumer<P> onApplicationStarted);
-
-		/**
-		 * @param frame the supplies the frame to use
-		 * @return this Builder instance
-		 */
-		Builder<M, P> frame(Supplier<JFrame> frame);
-
-		/**
-		 * @param frameTitle the frame title
-		 * @return this Builder instance
-		 */
-		Builder<M, P> frameTitle(String frameTitle);
-
-		/**
-		 * For a dynamic frame title.
-		 * @param frameTitle the observable controlling the frame title
-		 * @return this Builder instance
-		 */
-		Builder<M, P> frameTitle(Observable<String> frameTitle);
-
-		/**
-		 * @param mainMenu if true then a main menu is included
-		 * @return this Builder instance
-		 */
-		Builder<M, P> mainMenu(boolean mainMenu);
-
-		/**
-		 * @param maximizeFrame specifies whether the frame should be maximized or use its preferred size
-		 * @return this Builder instance
-		 */
-		Builder<M, P> maximizeFrame(boolean maximizeFrame);
-
-		/**
-		 * @param displayFrame specifies whether the frame should be displayed or left invisible
-		 * @return this Builder instance
-		 */
-		Builder<M, P> displayFrame(boolean displayFrame);
-
-		/**
-		 * Specifies whether to set the default uncaught exception handler when starting the application, true by default.
-		 * @param uncaughtExceptionHandler if true the default uncaught exception handler is set on application start
-		 * @return this Builder instance
-		 * @see Thread#setDefaultUncaughtExceptionHandler(Thread.UncaughtExceptionHandler)
-		 */
-		Builder<M, P> uncaughtExceptionHandler(boolean uncaughtExceptionHandler);
-
-		/**
-		 * @param startupDialog if true then a progress dialog is displayed while the application is being initialized
-		 * @return this Builder instance
-		 */
-		Builder<M, P> startupDialog(boolean startupDialog);
-
-		/**
-		 * @param frameSize the frame size when not maximized
-		 * @return this Builder instance
-		 */
-		Builder<M, P> frameSize(@Nullable Dimension frameSize);
-
-		/**
-		 * @param defaultFrameSize the default frame size when no previous size is available in user preferences
-		 * @return this Builder instance
-		 */
-		Builder<M, P> defaultFrameSize(@Nullable Dimension defaultFrameSize);
-
-		/**
-		 * Starts the application on the Event Dispatch Thread.
-		 */
-		void start();
-
-		/**
-		 * Starts the application.
-		 * @param onEventDispatchThread if true then startup is performed on the Event Dispatch Thread
-		 */
-		void start(boolean onEventDispatchThread);
 	}
 }